Semantic UI wrapper for Shiny

With this library it’s easy to wrap Shiny with Semantic UI components. Add a few simple lines of code and some CSS classes to give your UI a fresh, modern and highly interactive look.

Options

There are a number of global options that affect shiny.semantic as well as Shiny behavior.The options can be set globally with `options()`

shiny.custom.semantic.cdn (defaults to `NULL`)

This controls from where the css and javascripts will be downloaded.

shiny.custom.semantic (defaults to `NULL`)

This allows to set custom local path to semantic dependencies.

shiny.minified (defaults to `TRUE`)

Defines including JavaScript as a minified or un-minified file.

See also

Author

Maintainer: Jakub Nowicki opensource+kuba@appsilon.com

Authors:

Filip Stachura filip@appsilon.com

Dominik Krzeminski

Krystian Igras

Adam Forys

Paweł Przytuła pawel@appsilon.com

Jakub Chojna jakub.chojna@appsilon.com

Olga Mierzwa-Sulima olga@appsilon.com

Tymoteusz Makowski tymoteusz.makowski@appsilon.com

Other contributors:

Ashley Baldry [contributor]

Pedro Manuel Coutinho da Silva pedro@appsilon.com [contributor]

Kamil Żyła kamil@appsilon.com [contributor]

Rabii Bouhestine [contributor]

Federico Rivandeira federico@appsilon.com [contributor]

Appsilon Sp. z o.o. [copyright holder]